defeitos

+++ skaf/Screen/JanelaPrincipal.xaml.cs
@@ -71,14 +71,18 @@ namespace skaf
 
 
         void novidades() {
+            
             MessageBox.Show("Uma revisão foi feita no app! O que mudou?\n\n" +
                 "-Correção de bugs.\n" +
                 "-Atualização do layout de uniformes\n\n" +
                 "Se você perdeu as modificações/modelos que criou, por favor, informe em (emanuel.junior@ska.com.br)");
+
+
+
         }
 
         void carregarModelos() {
-            if (Properties.Settings.Default.primeiroLogin == true)
+            
                 novidades();
 
             conteiner.Children.Clear();
@@ -89,15 +93,21 @@ namespace skaf
                 {
                     DirectoryInfo a = new DirectoryInfo(Path.Combine(past.FullName,"Anexos"));
                     DirectoryInfo p = new DirectoryInfo(Path.Combine(past.FullName, "Admissões - automação"));
-                    a.Create();
-                    p.Create();
+                    DirectoryInfo p1 = new DirectoryInfo(Path.Combine(past.FullName, "Admissões - sistemas"));
+                    
+                    
+                  
+                   
 
-                    File.WriteAllBytes(Path.Combine(a.FullName, "Unimed Fesp Nacional - Apresentação.doc"),Properties.Resource1.Unimed_Fesp_Nacional___Apresentação);
-                    File.WriteAllBytes(Path.Combine(a.FullName, "Termo de Opção - Dental SKA.pdf"), Properties.Resource1.Termo_de_Opção___Dental_SKA);
-                    File.WriteAllBytes(Path.Combine(a.FullName, "planilha para inclusões- TIPAN.xls"), Properties.Resource1.planilha_para_inclusões__TIPAN);
-                    File.WriteAllBytes(Path.Combine(a.FullName, "Lista de procedimentos - Odonto.pdf"), Properties.Resource1.Lista_de_procedimentos___Odonto);
-                    File.WriteAllBytes(Path.Combine(a.FullName, "Formulário Designação de Beneficiários.pdf"), Properties.Resource1.Formulário_Designação_de_Beneficiários);
-
+                    if (!a.Exists)
+                    {
+                        a.Create();
+                        File.WriteAllBytes(Path.Combine(a.FullName, "Unimed Fesp Nacional - Apresentação.doc"), Properties.Resource1.Unimed_Fesp_Nacional___Apresentação);
+                        File.WriteAllBytes(Path.Combine(a.FullName, "Termo de Opção - Dental SKA.pdf"), Properties.Resource1.Termo_de_Opção___Dental_SKA);
+                        File.WriteAllBytes(Path.Combine(a.FullName, "planilha para inclusões- TIPAN.xls"), Properties.Resource1.planilha_para_inclusões__TIPAN);
+                        File.WriteAllBytes(Path.Combine(a.FullName, "Lista de procedimentos - Odonto.pdf"), Properties.Resource1.Lista_de_procedimentos___Odonto);
+                        File.WriteAllBytes(Path.Combine(a.FullName, "Formulário Designação de Beneficiários.pdf"), Properties.Resource1.Formulário_Designação_de_Beneficiários);
+                    }
                     string[] paths = {
                         Path.Combine(p.FullName, "Seguro de Vida.txt"),
                         Path.Combine(p.FullName, "Plano Unimed.txt"),
@@ -116,8 +126,13 @@ namespace skaf
 
                     };
 
+
+                    if (!p.Exists)
+                    {
+                        p.Create();
+                    
                     //cria arquivo dos e-mails
-                         File.WriteAllText(p + "/Gympass.txt", Properties.Resource1.Gympass);
+                    File.WriteAllText(p + "/Gympass.txt", Properties.Resource1.Gympass);
                         File.WriteAllText(Path.Combine(p.FullName, "Vale Transporte.txt"), Properties.Resource1.VALE_TRANSPORTE);
 
                     //att 1
@@ -143,10 +158,9 @@ namespace skaf
                         }
 
                     }
+                    }
 
 
-                    DirectoryInfo p1 = new DirectoryInfo(Path.Combine(past.FullName, "Admissões - sistemas"));
-                    p1.Create();
                     string[] paths1 = {
                         Path.Combine(p1.FullName, "Seguro de Vida.txt"),
                         Path.Combine(p1.FullName, "Plano Unimed.txt"),
@@ -166,6 +180,8 @@ namespace skaf
                     };
 
                     //cria arquivo dos e-mails
+                    if (!p1.Exists) { 
+                        p1.Create();
                     File.WriteAllText(p1 + "/Gympass.txt", Properties.Resource1.Gympass);
                     File.WriteAllText(Path.Combine(p1.FullName, "Vale Transporte.txt"), Properties.Resource1.VALE_TRANSPORTE);
 
@@ -195,7 +211,7 @@ namespace skaf
 
                     }
 
-
+                    }
